requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
daphne==4.1.2
dash==2.9.3
dash_ag_grid==31.2.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-sql-connector==3.4.0
Django==5.1.1
django-plotly-dash==2.3.1
django-redis==5.4.0
djangorestframework==3.15.2
dpd-components==0.1.0
et-xmlfile==1.1.0
Faker==30.0.0
Flask==3.0.3
fonttools==4.54.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib==3.9.2
msgpack==1.1.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pip==24.2
plotly==5.24.1
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.1.0
requests==2.32.3
retrying==1.3.4
seaborn==0.13.2
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
thrift==0.20.0
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0
zipp==3.20.2
zope.interface==7.0.3
